third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbfactory_deleteDatabase2-expected.txt

This is a testharness.js-based test.
PASS IDBFactory.deleteDatabase() - result of the request is set to undefined 
Harness: the test ran to completion.
